Split decoder and encoder into different files
