+++ src/codecs/png/decoder.rs
@@ -1,20 +1,13 @@
-//! Decoding and Encoding of PNG Images
-//!
-//! PNG (Portable Network Graphics) is an image format that supports lossless compression.
-//!
-//! # Related Links
-//! * <http://www.w3.org/TR/PNG/> - The PNG Specification
 use core::num::NonZeroU32;
-use std::borrow::Cow;
-use std::io::{BufRead, Seek, Write};
+use std::io::{BufRead, Seek};
 
-use png::{BlendOp, DeflateCompression, DisposeOp};
+use png::{BlendOp, DisposeOp};
 
 use crate::animation::Delay;
 use crate::color::{ColorType, ExtendedColorType};
 use crate::error::{
-    DecodingError, EncodingError, ImageError, ImageResult, LimitError, LimitErrorKind,
-    ParameterError, ParameterErrorKind, UnsupportedError, UnsupportedErrorKind,
+    DecodingError, ImageError, ImageResult, LimitError, LimitErrorKind, ParameterError,
+    ParameterErrorKind, UnsupportedError, UnsupportedErrorKind,
 };
 use crate::io::decoder::DecodedMetadataHint;
 use crate::io::{
@@ -23,16 +16,11 @@ use crate::io::{
 };
 use crate::math::Rect;
 use crate::metadata::LoopCount;
-use crate::utils::vec_try_with_capacity;
 use crate::{
-    DynamicImage, GenericImage, GenericImageView, ImageDecoder, ImageEncoder, ImageFormat,
-    ImageLayout, Limits, Luma, LumaA, Rgb, Rgba,
+    DynamicImage, GenericImage, GenericImageView, ImageDecoder, ImageFormat, ImageLayout, Limits,
+    Luma, LumaA, Rgb, Rgba,
 };
 
-// http://www.w3.org/TR/PNG-Structure.html
-// The first eight bytes of a PNG file always contain the following (decimal) values:
-pub(crate) const PNG_SIGNATURE: [u8; 8] = [137, 80, 78, 71, 13, 10, 26, 10];
-const XMP_KEY: &str = "XML:com.adobe.xmp";
 const IPTC_KEYS: &[&str] = &["Raw profile type iptc", "Raw profile type 8bim"];
 
 /// PNG decoder
@@ -322,7 +310,7 @@ impl<R: BufRead + Seek> ImageDecoder for PngDecoder<R> {
             .info()
             .utf8_text
             .iter()
-            .find(|chunk| chunk.keyword == XMP_KEY)
+            .find(|chunk| chunk.keyword == super::XMP_KEY)
             .cloned()
         {
             itx_chunk
@@ -687,274 +675,6 @@ impl<R: BufRead + Seek> ImageDecoder for ApngDecoder<R> {
     }
 }
 
-/// PNG encoder
-pub struct PngEncoder<W: Write> {
-    w: W,
-    compression: CompressionType,
-    filter: FilterType,
-    icc_profile: Vec<u8>,
-    exif_metadata: Vec<u8>,
-    xmp_metadata: Option<String>,
-}
-
-/// DEFLATE compression level of a PNG encoder. The default setting is `Fast`.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-#[non_exhaustive]
-#[derive(Default)]
-pub enum CompressionType {
-    /// No compression whatsoever
-    Uncompressed,
-    /// Fast, minimal compression
-    #[default]
-    Fast,
-    /// Balance between speed and compression level
-    Balanced,
-    /// High compression level
-    Best,
-    /// Detailed compression level between 1 and 9
-    Level(u8),
-}
-
-/// Filter algorithms used to process image data to improve compression.
-///
-/// The default filter is `Adaptive`.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-#[non_exhaustive]
-#[derive(Default)]
-pub enum FilterType {
-    /// No processing done, best used for low bit depth grayscale or data with a
-    /// low color count
-    NoFilter,
-    /// Filters based on previous pixel in the same scanline
-    Sub,
-    /// Filters based on the scanline above
-    Up,
-    /// Filters based on the average of left and right neighbor pixels
-    Avg,
-    /// Algorithm that takes into account the left, upper left, and above pixels
-    Paeth,
-    /// Uses a heuristic to select one of the preceding filters for each
-    /// scanline rather than one filter for the entire image
-    #[default]
-    Adaptive,
-}
-
-impl<W: Write> PngEncoder<W> {
-    /// Create a new encoder that writes its output to ```w```
-    pub fn new(w: W) -> PngEncoder<W> {
-        PngEncoder {
-            w,
-            compression: CompressionType::default(),
-            filter: FilterType::default(),
-            icc_profile: Vec::new(),
-            exif_metadata: Vec::new(),
-            xmp_metadata: None,
-        }
-    }
-
-    /// Create a new encoder that writes its output to `w` with `CompressionType` `compression` and
-    /// `FilterType` `filter`.
-    ///
-    /// It is best to view the options as a _hint_ to the implementation on the smallest or fastest
-    /// option for encoding a particular image. That is, using options that map directly to a PNG
-    /// image parameter will use this parameter where possible. But variants that have no direct
-    /// mapping may be interpreted differently in minor versions. The exact output is expressly
-    /// __not__ part of the SemVer stability guarantee.
-    ///
-    /// Note that it is not optimal to use a single filter type, so an adaptive
-    /// filter type is selected as the default. The filter which best minimizes
-    /// file size may change with the type of compression used.
-    pub fn new_with_quality(
-        w: W,
-        compression: CompressionType,
-        filter: FilterType,
-    ) -> PngEncoder<W> {
-        PngEncoder {
-            w,
-            compression,
-            filter,
-            icc_profile: Vec::new(),
-            exif_metadata: Vec::new(),
-            xmp_metadata: None,
-        }
-    }
-
-    fn encode_inner(
-        self,
-        data: &[u8],
-        width: u32,
-        height: u32,
-        color: ExtendedColorType,
-    ) -> ImageResult<()> {
-        let (ct, bits) = match color {
-            ExtendedColorType::L8 => (png::ColorType::Grayscale, png::BitDepth::Eight),
-            ExtendedColorType::L16 => (png::ColorType::Grayscale, png::BitDepth::Sixteen),
-            ExtendedColorType::La8 => (png::ColorType::GrayscaleAlpha, png::BitDepth::Eight),
-            ExtendedColorType::La16 => (png::ColorType::GrayscaleAlpha, png::BitDepth::Sixteen),
-            ExtendedColorType::Rgb8 => (png::ColorType::Rgb, png::BitDepth::Eight),
-            ExtendedColorType::Rgb16 => (png::ColorType::Rgb, png::BitDepth::Sixteen),
-            ExtendedColorType::Rgba8 => (png::ColorType::Rgba, png::BitDepth::Eight),
-            ExtendedColorType::Rgba16 => (png::ColorType::Rgba, png::BitDepth::Sixteen),
-            _ => {
-                return Err(ImageError::Unsupported(
-                    UnsupportedError::from_format_and_kind(
-                        ImageFormat::Png.into(),
-                        UnsupportedErrorKind::Color(color),
-                    ),
-                ))
-            }
-        };
-
-        let comp = match self.compression {
-            CompressionType::Balanced => png::Compression::Balanced,
-            CompressionType::Best => png::Compression::High,
-            CompressionType::Fast => png::Compression::Fast,
-            CompressionType::Uncompressed => png::Compression::NoCompression,
-            CompressionType::Level(0) => png::Compression::NoCompression,
-            CompressionType::Level(_) => png::Compression::Fast, // whatever, will be overridden
-        };
-
-        let advanced_comp = match self.compression {
-            // Do not set level 0 as a Zlib level to avoid Zlib backend variance.
-            // For example, in miniz_oxide level 0 is very slow.
-            CompressionType::Level(n @ 1..) => Some(DeflateCompression::Level(n)),
-            _ => None,
-        };
-
-        let filter = match self.filter {
-            FilterType::NoFilter => png::Filter::NoFilter,
-            FilterType::Sub => png::Filter::Sub,
-            FilterType::Up => png::Filter::Up,
-            FilterType::Avg => png::Filter::Avg,
-            FilterType::Paeth => png::Filter::Paeth,
-            FilterType::Adaptive => png::Filter::Adaptive,
-        };
-
-        let mut info = png::Info::with_size(width, height);
-
-        if !self.icc_profile.is_empty() {
-            info.icc_profile = Some(Cow::Borrowed(&self.icc_profile));
-        }
-        if !self.exif_metadata.is_empty() {
-            info.exif_metadata = Some(Cow::Borrowed(&self.exif_metadata));
-        }
-
-        let mut encoder =
-            png::Encoder::with_info(self.w, info).map_err(ImageError::from_png_enc)?;
-
-        if let Some(xmp_text) = self.xmp_metadata {
-            encoder
-                .add_itxt_chunk(XMP_KEY.to_string(), xmp_text)
-                .map_err(ImageError::from_png_enc)?;
-        }
-
-        encoder.set_color(ct);
-        encoder.set_depth(bits);
-        encoder.set_compression(comp);
-        if let Some(compression) = advanced_comp {
-            encoder.set_deflate_compression(compression);
-        }
-        encoder.set_filter(filter);
-        let mut writer = encoder.write_header().map_err(ImageError::from_png_enc)?;
-        writer
-            .write_image_data(data)
-            .map_err(ImageError::from_png_enc)?;
-        writer.finish().map_err(ImageError::from_png_enc)
-    }
-}
-
-impl<W: Write> ImageEncoder for PngEncoder<W> {
-    /// Write a PNG image with the specified width, height, and color type.
-    ///
-    /// For color types with 16-bit per channel or larger, the contents of `buf` should be in
-    /// native endian. `PngEncoder` will automatically convert to big endian as required by the
-    /// underlying PNG format.
-    #[track_caller]
-    fn write_image(
-        self,
-        buf: &[u8],
-        width: u32,
-        height: u32,
-        color_type: ExtendedColorType,
-    ) -> ImageResult<()> {
-        use ExtendedColorType::*;
-
-        let expected_buffer_len = color_type.buffer_size(width, height);
-        assert_eq!(
-            expected_buffer_len,
-            buf.len() as u64,
-            "Invalid buffer length: expected {expected_buffer_len} got {} for {width}x{height} image",
-            buf.len(),
-        );
-
-        // PNG images are big endian. For 16 bit per channel and larger types,
-        // the buffer may need to be reordered to big endian per the
-        // contract of `write_image`.
-        // TODO: assumes equal channel bit depth.
-        match color_type {
-            L8 | La8 | Rgb8 | Rgba8 => {
-                // No reodering necessary for u8
-                self.encode_inner(buf, width, height, color_type)
-            }
-            L16 | La16 | Rgb16 | Rgba16 => {
-                // Because the buffer is immutable and the PNG encoder does not
-                // yet take Write/Read traits, create a temporary buffer for
-                // big endian reordering.
-                let mut reordered;
-                let buf = if cfg!(target_endian = "little") {
-                    reordered = vec_try_with_capacity(buf.len())?;
-                    reordered.extend(buf.as_chunks::<2>().0.iter().flat_map(|le| [le[1], le[0]]));
-                    &reordered
-                } else {
-                    buf
-                };
-                self.encode_inner(buf, width, height, color_type)
-            }
-            _ => Err(ImageError::Unsupported(
-                UnsupportedError::from_format_and_kind(
-                    ImageFormat::Png.into(),
-                    UnsupportedErrorKind::Color(color_type),
-                ),
-            )),
-        }
-    }
-
-    fn set_icc_profile(&mut self, icc_profile: Vec<u8>) -> Result<(), UnsupportedError> {
-        self.icc_profile = icc_profile;
-        Ok(())
-    }
-
-    fn set_exif_metadata(&mut self, exif: Vec<u8>) -> Result<(), UnsupportedError> {
-        self.exif_metadata = exif;
-        Ok(())
-    }
-
-    fn set_xmp_metadata(&mut self, xmp: Vec<u8>) -> Result<(), UnsupportedError> {
-        self.xmp_metadata = Some(String::from_utf8(xmp).map_err(|_| {
-            UnsupportedError::from_format_and_kind(
-                ImageFormat::Png.into(),
-                UnsupportedErrorKind::GenericFeature("XMP metadata is not valid UTF-8".to_string()),
-            )
-        })?);
-        Ok(())
-    }
-
-    fn make_compatible_img(
-        &self,
-        _: crate::io::encoder::MethodSealedToImage,
-        img: &DynamicImage,
-    ) -> Option<DynamicImage> {
-        use ColorType::*;
-        match img.color() {
-            L32F => Some(img.to_luma16().into()),
-            La32F => Some(img.to_luma_alpha16().into()),
-            Rgb32F => Some(img.to_rgb16().into()),
-            Rgba32F => Some(img.to_rgba16().into()),
-            L8 | La8 | Rgb8 | Rgba8 | L16 | La16 | Rgb16 | Rgba16 => None,
-        }
-    }
-}
-
 impl ImageError {
     fn from_png(err: png::DecodingError) -> ImageError {
         use png::DecodingError::*;
@@ -974,16 +694,6 @@ impl ImageError {
             LimitsExceeded => {
                 ImageError::Limits(LimitError::from_kind(LimitErrorKind::InsufficientMemory))
             }
-        }
-    }
-    fn from_png_enc(err: png::EncodingError) -> ImageError {
-        use png::EncodingError::*;
-        match err {
-            IoError(error) => ImageError::IoError(error),
-            LimitsExceeded => {
-                ImageError::Limits(LimitError::from_kind(LimitErrorKind::InsufficientMemory))
-            }
-            _ => ImageError::Encoding(EncodingError::new(ImageFormat::Png.into(), Box::new(err))),
         }
     }
 }
@@ -1054,7 +764,7 @@ fn blend_pixel_bytes(bytes: &mut [u8], layout: &ImageLayout, from: &[u8], region
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::io::free_functions::decoder_to_vec;
+    use crate::{io::free_functions::decoder_to_vec, ImageEncoder};
     use std::io::{BufReader, Cursor, Read};
 
     #[test]
@@ -1120,7 +830,7 @@ mod tests {
 
         let mut encoded = Vec::new();
         {
-            let mut encoder = PngEncoder::new(&mut encoded);
+            let mut encoder = super::super::PngEncoder::new(&mut encoded);
             encoder.set_xmp_metadata(xmp.clone()).unwrap();
             encoder
                 .write_image(&img, 3, 1, ExtendedColorType::Rgb8)
